Conversion Rate Prediction
The data revolution has a lot to do with the fact that now we are able to collect all sorts of data about people who buy something on our site as well as people who don't. This gives us a tremendous opportunity to understand what's working well (and potentially scale it even further) and what's not working well (and fix it). 
The goal of this challenge is to build a model that predicts conversion rate and, based on the model, come up with ideas to improve revenue.